pyproject.toml
[project]
name = "medical-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask==3.1.1",
    "huggingface-hub==0.27.1",
    "langchain==0.3.26",
    "langchain-community==0.3.26",
    "langchain-groq>=0.3.8",
    "langchain-huggingface==0.1.2",
    "langchain-pinecone==0.2.8",
    "pypdf==5.6.1",
    "python-dotenv==1.1.0",
    "sentence-transformers==4.1.0",
]
